Comments are deleting.
Current issue: won't allow comments to be created one after the other. The new comment is replacing the old one.
I reload the page and already had a comment on the page, "comment1". I make a new comment, "comment2". If I delete comment1, comment2 deletes too...
when I reload the page, comment2 is there but comment is deleted.