Comments are submitting and being deleted correctly.
RESOLVED: ISSUE: won't allow comments to be created one after the other. The new comment is replacing the old one.I reload the page and already had a comment on the page, "comment1". I make a new comment, "comment2". If I delete comment1, comment2 deletes too. When I reload the page, comment2 is there but comment is deleted.

-SOLUTION: The old div was being deleted/replaced onclick while traversing through the dom. I made a separate div underneath it which would be targeted correctly via a special th:attribute. That way the old div only handled the prepending ajax html while the new div wrapped around the comment and gave the comment a unique id. I also took the previous html file I was using for ajax and placed it in a fragment and allowed my controller method to return that instead, more efficient that way and doesn't mess with Thymeleaf, sec:authorize, and CSS.